requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
google-genai==1.5.0
anthropic>=0.39.0
python-dotenv==1.0.1
psutil==6.1.1
pyautogui==0.9.54
pillow==11.1.0
python-multipart==0.0.20
websockets==14.2
